Add helper for fix

+++ src/support/converter_utils.cpp
@@ -7,7 +7,7 @@
 
 void OutputStats(const std::string& stat_name, double converted_value, double base_value)
 {
-   double percent = static_cast<double>(converted_value - base_value) / static_cast<double>(base_value) * 100.0;
+   double percent = (converted_value - base_value) / base_value * 100.0;
    if (!std::isfinite(percent))
    {
       percent = 0.0;
